backend/requirements.txt
aerich==0.8.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
asyncclick==8.1.7.2
asyncpg==0.30.0
click==8.1.8
dictdiffer==0.9.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
pyasn1==0.6.1
pydantic==2.10.4
pydantic_core==2.27.2
pypika-tortoise==0.3.2
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
PyYAML==6.0.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tomlkit==0.13.2
tortoise-orm==0.23.0
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
httpx==0.28.1